change type of id query

+++ src/main/java/br/com/dea/management/user/service/UserService.java
@@ -27,7 +27,7 @@ public class UserService {
         return this.userRepository.findByLinkedin(linkedin).orElseThrow(() -> new NotFoundException(User.class, linkedin));
     }
 
-    public User findById(long id) {
+    public User findById(int id) {
         return this.userRepository.findById(id).orElseThrow(() -> new NotFoundException(User.class, id));
     }
 }
